Modified button.js to implement CLONE and DELETE. Fixes #193.
Syntax is as in @RaphaelAlvez comment:
```JSON
{
  "func": "CLONE",
  "source": "myCollection", (or "widgets": "myCollection"),
  "count": <number>,
  "properties": {"applyVariables": [
     {
         "parameter": "<existing widget property>",
         "variable": "<new value>"
     },
    {
        "parameter": "<new, custom property of this set of clones>",
        "variable": "<value>"
    }
  ],
  "customProperty":"anotherValue"}
  "collection": "myCloneCollection"
```

In addition, a `clonedFrom` property is added to each cloned widget containing the id of the widget from which it was cloned.

Note that in this implementation, children of the cloned widget are not also cloned.

A small example is at http://212.47.248.129:3261/clones