place value compression of repeating color values (#499)
This uses `#$%&'()` to encode the repeat length for most colors. The color character is encoded first, then the code for how many times it is repeated if the color occurs for at least 3 pixels in a row (a color occurring only twice in a row the color code simply gets repeated). The characters `*+,-./0` encode the repeat length of the color `0`; but `0` is not always in the encoding. To make `0` mean one `0` the values of the characters used are 1-7 with the highest code point representing the lowest value (no zero is needed, 14 is encoded as 7 ones + 1 seven, 56 is 7 ones + 7 sevens). For the maximum supported canvas size (500x500 - 2500 pixels per region) the maximum possible length of a repeat encoding is 4 characters (7 + 7 x 7 + 7 x 7^2 + 7 x 7^3 = 2800). For a 200x200 canvas (400 pixels per region) 3 character repeat codes are always sufficient (7 + 7 x 7 + 7 x 7^2 = 399).