Log scale & 0.5 default
This converts the volume on the slider to a log scale and allows the user to set a volume of approximately twice the decibels as the new default value.  We experience volume on a log scale so a sound that is actually 10 times as intense appears to be twice as loud.  This change makes the slider change decibels rather than intensity so the volume will go from zero to about twice (in decibels).