Edit Mode converts width and height to float value
This is to address issue #708

This is a fairly simple solution.  It will convert all string numbers for width and height to float values.  I think that is what they should always be anyway, but if I am wrong, then that is a problem with this PR.

Non-number values will also continue to default to 800 for width and 500 for height.  I believe this is because it is half of the range for each.  Personally, I don't see that as much of an issue since it is better than it entering a word as the width or height.