Remove width/height equal to cardDefault/cardType
This PR should address issue #680 and remove the width and height if those values are already being inherited from either cardDefault or cardType.  So, if it is a value that is overriding the cardDefault or cardType it will remain on the card, but if it is not, then it will be removed.