Edit Mode converts width and height to float value (#709)
* Edit Mode converts width and height to float value

This is to address issue #708

This is a fairly simple solution.  It will convert all string numbers for width and height to float values.  I think that is what they should always be anyway, but if I am wrong, then that is a problem with this PR.

Non-number values will also continue to default to 800 for width and 500 for height.  I believe this is because it is half of the range for each.  Personally, I don't see that as much of an issue since it is better than it entering a word as the width or height.

* Updated to get width and height from text input

* Conditional to reject non-numbers

* Only updates width/height if valid number

* More concise Edit Mode Adjustment (#710)

This should do the exact same thing as PR #709, however, I realized this would be a more concise way of doing the same thing since it was repetitive before.