Close loopholes in dropLimit and enterRoutine
This is intended to address issue #183.  Hopefully you should not be able to circumvent the drop limit while a card is being dragged from the holder.

Since it was related, it also addresses the loophole in enterRoutine collecting oldParentID when a widget is dragged.

enterRoutine: An enterRoutine will be called whenever a new child gets added to the widget it is in. The routine will have the preset variable oldParentID, referring to the previous parent of the child, and the preset collection child, containing the id of the new child. (Note that if the parent changes as a result of a programmatic action, this will work as [likely] expected. But if the parent changes as the result of a user dragging the widget, the old parent will be lost when the widget is "picked up", so that oldParentID will be null. One way around this problem is to add a parentChangeRoutine to the widget which stores a non-null old parent somewhere that the enterRoutine can find it.)