Add arrays to dropTarget
This is to address issue #775.  The idea is to allow an array for dropTarget values.  Then if the widget meets one of the values, it can be dropped into a holder as long as it meets all the other criteria.

Example:
```json
"dropTarget": {
  "deck": ["deck1", "deck2"],
  "cardType": "thisCardType"
}
```
Should allow a card with cardType `thisCardType` to be dropped in the holder as long as it was from `deck1` or `deck2`.