Add array of objects to dropTarget (#777)
* Add arrays to dropTarget

This is to address issue #775.  The idea is to allow an array for dropTarget values.  Then if the widget meets one of the values, it can be dropped into a holder as long as it meets all the other criteria.

Example:
```json
"dropTarget": {
  "deck": ["deck1", "deck2"],
  "cardType": "thisCardType"
}
```
Should allow a card with cardType `thisCardType` to be dropped in the holder as long as it was from `deck1` or `deck2`.

* Excluded fixedParent & removed unneeded part

* Removed fixedParent b/c doesn't work

* Update main.js

* Update to allow array of objects

* Update holder.js

* Update holder.js

* Update main.js

* changed array helper fuction to match PR #788

* Newline

* Update fileupdater.mjs

* remove duplicate

* Update fileupdater.mjs

* var

* Update fileupdater.mjs

hoping to paste over typo

* Trying let

* trying const

* Getting rid of unnecessary declaration

* Update childrenFilter

So function will work properly on holder children

* Update fileupdater.mjs

* Update fileupdater.mjs

* Update main.js

* Update holder.js

* Update holder.js

* Update main.js

* Update holder.js

* Update main.js

* change property to dropTarget

* Update main.js

* Update main.js

* Update main.js

Co-authored-by: David Robarts <368796+robartsd@users.noreply.github.com>
Co-authored-by: ArnoldSmith86 <arnold.smith.86@mail.ru>