Primarily, refactor onClickUpdateWidget.
Change #addCustomWidgetOverlay to #addBasicWidget. Also refactor onClickUpdateWidget to handle separately the cases where it is called from the basic editor and where it is called from the JSON editor; the difference is whether showOverlay() is called or not. Finally, fix (again) jeAddWidgetDone so that it shows the context actions for the newly-added widget.