Reuse edit mode add widget screen in JSON editor (#817)
* Change 'add widget' code in jsonedit.js to simply show the add overlay.
Modify addWidgetLocal to return the id of the added widget, and use that id in various places in editmode.js to properly configure the JSON editor when the add overlay is done.

* Modified calls to addWidgetLocal in most cases to used returned widget id.

* Cause the id of the newly created widget to be highlighted in the JSON editor.

* Changed "Add custom widget" to "Add basic widget", and modified code to indeed add an empty basic widget at (0,0), with text "basic widget".
Also modified the layout of the "Interactive" section of the edit overlay just for appearance.
Finally, moved some code around in editmode.js so that order of widget creation in the Interactive section corresponds to appearance on the screen.

* Tried to fix TestCafe. Failed.

* Changed dynamic expressions testcafe test. Still doesn't work.

* fixed TestCafe by replacing NO-BREAK SPACE by normal spaces and the fact that there is an ID generation before the button execution now means that the random numbers are different

* Update jsonedit.js

Change two occurrences of `jeGetEditorContent` to `jeSetEditorContent`

* fixed typos

* removed addCustomOverlay

* removed CSS for addCustomOverlay

* Fix CSS for buttons in editmode.css
Remove unneeded calls to json editor in overlayDone

* Another modification to overlayDone. Still not quite correct.

* Add call to jeApplyChanges to work around updating issue.

* Copied onClickUpdateWidget from editmode.js to jsonedit.js and modified it for use in new widget creation. Notably, removed the trailing call to showOverlay().

* Moved JSON editor-specific code in overlayDone into jsonedit.js, and simplified that code as originally suggested. Without the extra call to showOverlay in onClickUpdateWidget (now replaced by jeUpdateWidget) this all seems to work.

* Re-insert call to jeApplyChanges.

* Added a few comments to editmode.js

* Primarily, refactor onClickUpdateWidget.

Change #addCustomWidgetOverlay to #addBasicWidget. Also refactor onClickUpdateWidget to handle separately the cases where it is called from the basic editor and where it is called from the JSON editor; the difference is whether showOverlay() is called or not. Finally, fix (again) jeAddWidgetDone so that it shows the context actions for the newly-added widget.

* Same as last commit message.

* Fix testcafe tests.

* Modify button css in add overlay to put buttons on a line by themselves. Move other widgets around as needed to accommodate.

* removed unnecessary variable

* Revert removal of temporary constant

Co-authored-by: ArnoldSmith86 <arnold.smith.86@mail.ru>