Update input overlay widget updates
Adds the min and max to the input label.
If the number is greater than the max, it is adjusted to the max.
If the number is greater than the min, it is adjusted to the min.

Bonus fix: If a header is specified, it is inserted as a title.