Add gridAlign property which sets grid alignment for widgets.
This adds a new gridAlign property which allows specifying the point on
a widget to align to the grid point. By default it is now the top left.
This also fixes a bug in the existing grid code such that widgets are
now aligned correctly by the specified point and uses a fileupdater to
ensure no change is observed in existing games.