UI Explanation Update for In-Progress Games
This adds additional explanatory text to the UI when a user clicks on Save Active Game. Recently at least one new user has been confused by the use of that button and attempted to use it for game development, which causes some problems because of the additional metadata added. Ideally, we could figure out some better way to do this on the graphics and buttons before the player reaches this point, but I after thinking about it, I don't have any great ideas so this can be an interim solution.

The one idea I did have is an on-hover pop up over Save Active Game which would basically say the same thing as this modified text. We don't need tooltips on every button, but maybe just that one.