PL-Update Great Greece
Allows for full color customization of all components (including the turn cards), it also puts the turn order "on rails" with a forward and backward turn button. It also implements a counter system so that you can draw the type of track tile you want without having to flip it over, but it still counts the total amount.