Formalize use of dropTarget and dropLimit on non-holder widgets
As recently mentioned in Discord, it is possible to allow cards or any other widget to become a parent just by including dropLimit and dropTarget properties on the widgets that will be the parents. This just adds the default null value in widget.js so the JSON Editor recognizes that and recolors the properties yellow.