Add prompt before overwriting widgets with same id from buffer
Fixes #1845.  When you load widgets from the buffer into the game, it checks for widgets with the same id.  If widgets with the same id exist, you get a prompt with a list of the widgets that will be overwritten and need to press OK to continue or cancel to abort loading.