Fix crash when when SETting parent to self of descendant
Fixes #1414 for both routine operations and invalid typed entries in the JSON Editor.

There is a quirk with the editor that I cannot figure out.  Instead of getting red warning text, you get an alert pop-up with the warning.  When the pop-up is closed, the invalid parent still shows up in the editor, but it has not been applied.

I tried changing the jeGetContext function in jsonedit.js in the try block. But there were multiple problems with everything I tried.

What is currently in this code is better than what we have and it works as is to prevent crashes and provide warning. But it could be improved.