Update JSON Editor for count parameters
PR #1466 allowed "all" to be an option for the count parameter of FLIP, MOVE, MOVEXY, and ROTATE.  And PR #2377 allowed "loop" for AUDIO.  Neither of those PRs updated the JSON editor.  This fixes that.

It is imperfect because jeAddEnumCommands and jeAddNumberCommand functions do not exactly work well together and create two separate sections of buttons.  But this PR is a pretty good compromise that gives an option to choose all/loop or a number and doesn't require any changes to the underlying functions.