Interact with widgets without clicking
This is a proof of concept / demo.  I'm sure there are problems with it and things that can be improved.  See the test room.  Click on the blue button to rotate through 3 settings: select, deselect, or off.  You can also use the spacebar as a hotkey to cycle through the settings.  Move over the area of squares to turn them black or white.  Move over the cards in the hand to raise them up a few pixels.

Lots of uses for this, but maybe it is computationally expensive.  I don't know.  If implemented, it would be a good idea to have a permanent button where we have sound and lights to change the cursor status and maybe a permanent hot key since some uses require the deselect mode to be active while the cursor is in position.

Also, would be nice to expand this to an area select (click and drag a rectangle).

Have not tested on mobile.  Pretty sure it will not work in its current configuration, but I think it could be made to work.