Replace Google Material Icons with Material Symbols
This impacts games and the UI.

Needs a file updater and some visual testing of the UI because that stuff is everywhere.  I think I got everything.  But this was a bigger undertaking than expected because the default for the old Icons was "filled" and that had to be manually applied in some location in the UI.