Public Library - Update CONTEST
Removed "Recall" button for each player and condensed to a "New Game" button that recalls all tiles to holders, including Neutral discs, and resets score to 0.
Modified "Discard" button to discard all values except for 0 which will be returned to owners hand.
Moved Discard holder to bottom of play area instead of off screen to accommodate events when people discard accidentally.
Labelled Discard and Neutral discs to avoid any confusion regarding what they are for.
Placed pawn at the start position for the game (top of board in the 4 column). Added reset routine to revert pawn to start space.
Updated Metadata for "How to use" to address changes