Remove unneeded suit information in custom card decks
When adding a custom deck through the Properties toolbar, every deck gets detailed information about the position of the small suit icons.  For example for a 5, you get:

```
"cardType [5 of diamonds] (in deck)": {
    "suit": "/i/game-icons.net/skoll/diamonds.svg",
    "suitColor": "#e50932",
    "rank": 5,
    "suit-P11": "/i/game-icons.net/skoll/diamonds.svg",
    "suit-P13": "/i/game-icons.net/skoll/diamonds.svg",
    "suit-P51": "/i/game-icons.net/skoll/diamonds.svg",
    "suit-P53": "/i/game-icons.net/skoll/diamonds.svg",
    "suit-P32": "/i/game-icons.net/skoll/diamonds.svg"
  }
```
But the suit-Pnn info is only used on the standard deck. This PR prevents that additional and unnecessary info from being added to the cards in the other types of decks.