Tutorial - Update Holders
This removes the Moving: Basic and Moving: Advanced variants. They were really pretty simple and did not add much to the understanding of holders. They dated back to the earliest days of VTT before a lot of new capabilities essentially make everything in those variants either redundant or unnecessary to explain.

More importantly, this adds a new CSS Flex variant to described how to use CSS to make rows of widgets or compress widgets to make them fit into a holder.