Editor Displays SVG Colors (#2459)
* Editor Help with SVG Colors

Fixes #2454.  This is only a proof of concept.

I would like to have been able to get this to work with actual editor buttons, but that seemed impossible, so I settled on showing it in the JSON Editor like the duplicate widget button. I used that area as an example, but I could never get the div to display properly.  It does work in place of the duplicate widget button now as a demo. I could use some help getting the new div right.  The button is there, it just doesn't work as it should.  I've attached an image to show how this looks.  Pretty cool.

* Working code

* Add support for VTT icons with currentColor

* Close color viewer when widget is deselected.

* Update Icons tutorial, svgReplaces variant