Add random parameter to `SELECT` (#2462)
* Add random parameter to `SELECT`

SELECT always choose widgets in the same order.  Often I want a random selection but don't want to shuffle.  This PR adds a new parameter "random" that defaults to false so as not to change existing behavior. But if set to true, it uses a Fisher-Yates shuffle to select a random collection.  Of course, this only matters if the "max" is less than the total number of widgets in the collection.

* Reuse dom helper function

* Remove log statement

* Simplify call to shuffleArray

* Fix tutorial variant 2

I accidentally overwrote it.

* Move shuffle to after pile resolution

* Remove duplicate shuffle