Fix svg color picker in editor
The color picker added in #2459 only allowed the new icon button to work if the file was named .svg.  That means that the button would not work for uploaded files. This fixes that by checking the file contents.  It took some playing around to get the regex right.  Not all svg files are created equal.  It still may miss some, but this works on all the different ones I tried and excludes things that are obviously not svgs.