PL A Night on Town: Make personal and general money supplies more clear (#2539)
PL Game "A Night on' Town": User feedback was that it can be hard to determine that the visible coins are a personal supply vs. a general supply.

This update addresses this issue, both in-game and in the How to use this implementation section.

(Authored by Mitchell Stolycia as already credited in Attribution section.)