Public Library - Update Score Keeper
The original version was very difficult to use on mobile because the buttons were small.  This update reduces the number of possible players from 13 to 8 in order to increase the size of buttons and the score table.  Definitely worth the trade off as the number of games with more than 8 players is very small.