Public Library - Update Arrays Tutorial (#2573)
Adds a new Overview variant to explain basic and important concepts for arrays.  Updates the Creating variant.  It was poorly written when I barely understood arrays and now it is much better.  Minor cosmetic change to the Evaluating variant.  The changes to other variants were add automatically when I made the new Overview variant be the first one on the list.  Not sure why that happened, but they are the same.