Allow pan in zoomed edit mode with spacebar
This PR adds a check for the spacebar being held down in edit mode. If the spacebar is held down, then the multi-select rectangle is suppressed. If the zoom level > 1, and the spacebar is held down, then left-clicking the mouse and dragging enables panning no matter where the cursor is. This applies even over a movable widget because if the designer is holding down the spacebar, then they want to pan, not move widgets.