feat: Add DELAY routine command (#2702)
* feat: Add UPDATE routine command

This commit introduces a new UPDATE command to the client-side routine engine. This command allows routines to force a synchronization of the current game state with the server and then pause for a specified duration.

The UPDATE command accepts a wait parameter, which specifies the pause duration in milliseconds.

To ensure that pending changes are sent immediately, even when the routine is executed within a batch (atchStart/atchEnd), a new lushDelta function has been implemented. This function temporarily overrides the batching mechanism to send the current delta to the server.

The UPDATE command now performs the following actions:
1.  Calls lushDelta() to send all pending state changes.
2.  Pauses the routine's execution for the duration specified by the wait parameter.

This enables the creation of more complex, multi-step animations and actions that are visible to all players in real-time.

* refactor: Rename UPDATE to DELAY

* Added DELAY to json editor sidebar

* Fix wait vs time

This does not change the name, but it fixes a bug from the refactor

* Oops, missed an instance of wait

* Move DELAY to correct alphabetical location

* Restore delta cause between flushes

* fixed order

* renamed to milliseconds and changed to use sleep helper function

* should be enough to remember the delta cause inside flushDelta

* added debug summary

* added DELAY to validation

* update the text of the latest history entry when a new delta is received

* shortened a long line

---------

Co-authored-by: 96LawDawg <76912527+96LawDawg@users.noreply.github.com>
Co-authored-by: Arnold Smith <arnold.smith.86@mail.ru>